+++ versioned/tiered/tiered-tests/src/main/java/org/projectnessie/versioned/impl/SampleEntities.java
@@ -44,7 +44,7 @@ public class SampleEntities {
     return EntityType.L1.buildEntity(producer -> producer.commitMetadataId(createId(random))
         .children(IntStream.range(0, InternalL1.SIZE).mapToObj(x -> createId(random)))
         .ancestors(Stream.of(InternalL1.EMPTY.getId(), Id.EMPTY))
-        .keyMutations(Stream.of(Key.of(createString(random, 8), createString(random, 9)).asAddition((byte) 0)))
+        .keyMutations(Stream.of(Key.of(createString(random, 8), createString(random, 9)).asAddition(0)))
         .incrementalKeyList(InternalL1.EMPTY.getId(), 1));
   }
 
@@ -78,7 +78,7 @@ public class SampleEntities {
   public static InternalFragment createFragment(Random random) {
     return EntityType.KEY_FRAGMENT.buildEntity(producer -> producer.keys(IntStream.range(0, 10)
         .mapToObj(i -> Key.of(createString(random, 5), createString(random, 9), String.valueOf(i)))
-        .map(i -> WithEntityType.of((byte)0, i))));
+        .map(i -> WithEntityType.of(0, i))));
   }
 
   /**
@@ -105,7 +105,7 @@ public class SampleEntities {
               .unsaved()
               .delta(1, createId(random), createId(random))
               .mutations()
-              .keyMutation(Key.of(createString(random, 8), createString(random, 8)).asAddition((byte) 0))
+              .keyMutation(Key.of(createString(random, 8), createString(random, 8)).asAddition(0))
               .done();
         }));
   }
